+++ terraform/lambda_functions.tf
@@ -409,6 +409,7 @@ resource "aws_lambda_function" "send_to_gitc" {
       REGION                      = data.aws_region.current.name
       GIBS_REGION                 = var.gibs_region
       GIBS_SQS_URL                = "https://sqs.${var.gibs_region}.amazonaws.com/${var.gibs_account_id}/${var.gibs_queue_name}"
+      GIBS_RESPONSE_TOPIC_ARN     = aws_sns_topic.gibs_response_topic.arn
     }
   }