+++ tests/mock_gitc.py
@@ -3,6 +3,8 @@ from datetime import datetime, timezone
 from time import sleep
 from typing import Dict, List
 
+from moto import sqs
+
 
 class GITC:
     def __init__(self, incoming_queue, outgoing_topic):
@@ -18,17 +20,17 @@ class GITC:
     def __exit__(self, exception_type, exception_value, traceback):
         self._incoming_queue.purge()
 
-    def wait_for_messages(self, count: int = 1) -> List[Dict]:
+    def wait_for_messages(self, count: int = 1) -> List[sqs.models.Message]:
         while len(self._received_messages) < count:
             sleep(1)
         return self._received_messages[0:count]
 
     def process_messages(self):
         while True:
-            messages = self._incoming_queue.receive_messages(WaitTimeSeconds=10)
+            messages = self._incoming_queue.receive_messages(WaitTimeSeconds=10, MessageAttributeNames=["All"])
             for message in messages:
                 message_body = json.loads(message.body)
-                self._received_messages.append(message_body)
+                self._received_messages.append(message)
 
                 response = {
                     "provider": message_body["provider"],